requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrdict3==2.0.2
attrs==21.4.0
charset-normalizer==2.0.12
colored==1.4.4
frozenlist==1.3.0
Gooey==1.0.8.1
idna==3.3
multidict==6.0.2
numpy==1.21.6
Pillow==9.0.1
psutil==5.9.4
pygtrie==2.5.0
scrython==1.10.1
six==1.16.0
tqdm==4.63.0
typing_extensions==4.1.1
wxPython==4.2.0
yarl==1.7.2
